changed Quantity to decimal

+++ BlazorAppWithMediatrAndDapper.BLL/Models/OrderItem.cs
@@ -5,7 +5,7 @@ public class OrderItem
 	public int Id { get; private set; }
 	public int OrderId { get; private set; }
 	public string Name { get; private set; }
-	public uint Quantity { get; private set; }
+	public decimal Quantity { get; private set; }
 	public string Unit { get; private set; }
 
 	private OrderItem() { }
